CO: Move self configurator classes in subfolder
